Fix some incorrect piracy flagging
If the type of a method *argument* is a union, then if *any* of the
types in the union is foreign, the whole type should count as
foreign.

But when installing a method for a union, then it should only count
as foreign if *all* the types in the union are foreign.